defaultPasswordValidator unit test
